+++ Migrations/Code/LegacyClassesForIde.php
@@ -9,14 +9,6 @@ namespace {
 // Configuration
 namespace {
     /**
-     * @deprecated will be removed in oelib 4.0.0
-     */
-    class Tx_Oelib_ConfigCheck extends \OliverKlee\Oelib\Configuration\ConfigurationCheck
-    {
-
-    }
-
-    /**
      * @deprecated will be removed in oelib 5.0.0
      */
     class Configuration extends \OliverKlee\Oelib\Configuration\TypoScriptConfiguration